requirements.txt
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
